% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppearCount.R
\name{AppearCount}
\alias{AppearCount}
\title{Counr appearance times}
\usage{
AppearCount(x, y)
}
\arguments{
\item{x}{book number}

\item{y}{name of character}
}
\value{
dataframe for appearance times of character in each chapter
}
\description{
Count appearance times of character in each chapter
}
\details{
Add book number from gutenberg and find appearance times of character in each chapter
}
\author{
Yahui Zhang
}
