% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgclose.R
\name{avgclose}
\alias{avgclose}
\title{Average total values in Close column which is the price the stock closed.}
\usage{
avgclose()
}
\arguments{
\item{sum}{A calculation}
}
\value{
sum divided by rows
}
\description{
Average total values in Close column which is the price the stock closed.
}

