% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avghigh.R
\name{avghigh}
\alias{avghigh}
\title{Average total values in High column which is the highest price the stock reached.}
\usage{
avghigh()
}
\arguments{
\item{sum}{A calculation}
}
\value{
sum divided by rows
}
\description{
Average total values in High column which is the highest price the stock reached.
}

