% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avglow.R
\name{avglow}
\alias{avglow}
\title{Average total values in Low column which is the lowest price the stock reached.}
\usage{
avglow()
}
\arguments{
\item{sum}{A calculation}
}
\value{
sum divided by rows
}
\description{
Average total values in Low column which is the lowest price the stock reached.
}

