% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgopen.R
\name{avgopen}
\alias{avgopen}
\title{Average total values in Open column which is the price the stock opened.}
\usage{
avgopen()
}
\arguments{
\item{sum}{A calculation}
}
\value{
sum divided by rows
}
\description{
Average total values in Open column which is the price the stock opened.
}
\examples{

}

