% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binet.R
\name{binet}
\alias{binet}
\title{A grid chart for bigrams}
\usage{
binet(x)
}
\arguments{
\item{x}{book file data name}
}
\value{
bigrams of the word
}
\description{
We want to show the network of bigrams in this book
}
\author{
Yi Xu
}
