% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitable.R
\name{bitable}
\alias{bitable}
\title{bigram frequency table}
\usage{
bitable(x, y)
}
\arguments{
\item{x}{book file data name}

\item{y}{word}
}
\value{
bigrams of the word
}
\description{
We want to find bigrams of a specific word in this book, if you input a word,
you can get the word associated with it and also the frequency of it.
}
\author{
Yi Xu
}
