% Generated by roxygen2 (4.0.2): do not edit by hand
\name{breakthrough.norm}
\alias{breakthrough.norm}
\title{plot the breaktrhough curve with normalized time and concentration}
\usage{
breakthrough.norm(file, bed_volume = 12)
}
\arguments{
\item{file}{name of input file}

\item{bed_volume}{volume of resin column, default value is 12}
}
\value{
A table of raw data, and a breakthrough curve
}
\description{
Input an excel file and print the normalized breakthrough curve.
}

