% Generated by roxygen2 (4.0.2): do not edit by hand
\name{breakthrough}
\alias{breakthrough}
\title{plot the breaktrhough curve}
\usage{
breakthrough(file)
}
\arguments{
\item{file}{name of input file}
}
\value{
A table of raw data, and a breakthrough curve
}
\description{
Input an excel file and print the breakthrough curve.
}

