% Generated by roxygen2 (4.0.2): do not edit by hand
\name{capacity}
\alias{capacity}
\title{caculate ion-exchange capacity of a chemical}
\usage{
capacity(file, weight, element)
}
\arguments{
\item{file}{name of input file}

\item{weight}{weight of the resin}

\item{element}{element you are interested, options: Na, Mg, Ca, Ba, Ra, Sr}
}
\value{
A sentence with the name of chemical element and its capacity
}
\description{
Input an excel file, the weight of resin and which chemical element you are interested, and output the ion-exchange capacity.
}

