% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{cv.glm.logit.ridge}
\alias{cv.glm.logit.ridge}
\title{Cross validation for ridge logistic regression using design matrix and Y}
\usage{
cv.glm.logit.ridge(X, y, lambdas, fold = 10)
}
\arguments{
\item{X}{an n by p matrix containing the design matrix X, the predictors}

\item{y}{an n by 1 vector containing the outcome variable}

\item{lambdas}{an vector of lambda for the grid search of best lambda}

\item{fold}{a integer as the number of fold in the cross validation}
}
\value{
the best lambda for ridge logistic regression after grid search
}
\description{
this function use cross validation to identify the best lambda for ridge logistic regression
}
