% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Reads api token called redcapTokenName from the users' .Renviron file; queries redcapUrl to return the redcap Report redcapReportID; and return the contents as a tibble}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\value{
content as a tibble
}
\description{
Reads api token called redcapTokenName from the users' .Renviron file; queries redcapUrl to return the redcap Report redcapReportID; and return the contents as a tibble
}
\examples{
redcapTokenName = "6189879441F5C29A25245880677488BF"
redcapUrl = "https://redcap.emory.edu/api/"
redcapReportId = '46524'

downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
