% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emo_hist.R
\name{Emo}
\alias{Emo}
\title{analysis main sentiment}
\usage{
Emo(x)
}
\arguments{
\item{x}{book number}
}
\value{
histogram for main sentiment of each chapter
}
\description{
Analysis main sentiment of each chapter in the book
}
\details{
Add book number from gutenberg and make a histogram for main sentiment of each chapter in the book.
}
\author{
Yahui Zhang
}
