% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{New label for the leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{is a cleaned data set of the earthquakes}
}
\value{
a data frame with a column for the labels' content
}
\description{
eq_create_label creates the content of a label that includes the
locationm the magnitude and the number of deaths of the earthquake
}
\note{
if one of the information is missing the it will not appear in
the label
}
\examples{
\\code{eq \%>\% mutate(popup_text = eq_create_label(.)) \%>\%
             eq_map(annot_col = "popup_text")}
}
