% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Leaflet map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{is a cleaned data set of the earthquakes}

\item{annot_col}{is the character name of the collumn that the user
wants to appear in the pop-up window.}
}
\value{
a leaflet map
}
\description{
eq_map creates a leaflet map for the earthquakes in our data set.
}
\examples{
eq \%>\%
           filter((COUNTRY \%in\% c("GREECE")) & year(DATE)>1981) \%>\%
           eq_map(annot_col = "DEATHS")





}
