% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_table.R
\name{export_deg_excel}
\alias{export_deg_excel}
\title{Excel plot script}
\usage{
export_deg_excel(deg_table, out_xlsx = "DEG_results.xlsx")
}
\arguments{
\item{deg_table}{A data.frame  containing DEG results to export.}

\item{out_xlsx}{Output Excel filename or path}
}
\value{
The path/filename of the saved Excel file.
}
\description{
Writes a differential expression results table to an excel file using `openxlsx`.
The output contains one sheet named "DEG"
}
