% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_low_expressions.R
\name{filter_low_expression}
\alias{filter_low_expression}
\title{Filters lowly expressed genes and creates a normalised list using edgeR and DGElist
#´@description
This code aligns the sample metadata and the count matrix, also removes lowly expressed genes.}
\usage{
filter_low_expression(counts, samples, group_col = "disease")
}
\arguments{
\item{counts}{A data frame of raw counts}

\item{samples}{a data frame with sample table}

\item{group}{col a column names with in samples}
}
\value{
a list of filtered expression from samples
}
\description{
Filters lowly expressed genes and creates a normalised list using edgeR and DGElist
#´@description
This code aligns the sample metadata and the count matrix, also removes lowly expressed genes.
}
