% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Timeline of Earthquackes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
geom_timeline creates a diagram that involve a line with points that
indicate the earthquakes in chronological order.
}
\examples{
\\code{ggplot(data=eq \%>\% filter(COUNTRY == "GREECE" &
                year(DATE)>1970)) +
geom_timeline(aes(x=DATE, y =COUNTRY ,size = INTENSITY,color = DEATHS))+
theme(panel.grid = element_blank(),
     axis.ticks.y = element_blank(),
     axis.title.y = element_blank(),
     panel.background = element_blank())}

}
