% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\docType{data}
\name{Geomtimeline_label}
\alias{Geomtimeline_label}
\title{Labels for the Timeline of Earthquackes}
\format{An object of class \code{Geomtimeline_label} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
Geomtimeline_label
}
\description{
geom_timeline_label adds labels to the timeline of earthquakes. The user
can choose the number n_max of earthquakes  of greatest intensity for
which the labels wants to appear.
}
\keyword{datasets}
