% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{glm.logit.ridge}
\alias{glm.logit.ridge}
\title{ridge logistic regression using design matrix and Y}
\usage{
glm.logit.ridge(X, y, lambda = 1, tol = 1e-07, maxit = 50, trace = FALSE)
}
\arguments{
\item{X}{an n by p matrix containing the design matrix X, the predictors}

\item{y}{an n by 1 vector containing the outcome variable}

\item{lambda}{a scalar value as penalty for ridge logistic regression}
}
\value{
the IRLS estimate of regression coefficients
}
\description{
this function fit the ridge logistic regression using design matrix X and response variable Y
this model use iteratively reweighted least square algorithm to estimate coefficient of X
}
