% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts_new.R
\name{load_counts}
\alias{load_counts}
\title{Load RNA-seq count matrix from a tab-delimited file}
\usage{
load_counts(counts_path)
}
\description{
@description
 It reads a tab-separated counts table where the first column is gene identifiers.
}
\details{
@param counts_path It paths to the counts file.

 @return A numeric matrix of counts with genes as rownames and samples IDs as a colnames.
 @export
}
