% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_samples_new.R
\name{load_samples}
\alias{load_samples}
\title{Load samples metadata samples}
\usage{
load_samples(samples_path)
}
\arguments{
\item{sample_path}{It pats to the sample table}
}
\value{
A data.frame contains sample annotations.
}
\description{
It reads a tab-separated sample table and converts.
}
