% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{logLik}
\alias{logLik}
\title{log likelihood function for logistic regression}
\usage{
logLik(eta, y)
}
\arguments{
\item{eta}{an n by 1 vector containing the linear combination of predictors and coefficient}

\item{y}{an n by 1 vector containing the outcome variable}
}
\value{
the log likelihood of logistic regression
}
\description{
This function calculates the log likelihood of logistic regression
based on given vector of eta and outcome
}
