% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Takes a vector c and searches for the parameter p that maximizes the the log likelihood log(P(p|c)) using gridsearch with interval 0.001}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A binary vector.}
}
\value{
p that maximizes the log likelihood
}
\description{
Takes a vector c and searches for the parameter p that maximizes the the log likelihood log(P(p|c)) using gridsearch with interval 0.001
}
\examples{
logLikBernoulli(c(0,1,0,0,1))
}
