% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Takes either one (x1) or two (x2) samples from preliminary data and returns minimum sample size for the t test of the null hypothesis that either mean(x1) = 0 or mean(x1) = mean(x2) with 80% power at alpha = 0.05}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{numerical vector}

\item{x2}{(optional) numerical vector}
}
\value{
minimum sample size
}
\description{
Takes either one (x1) or two (x2) samples from preliminary data and returns minimum sample size for the t test of the null hypothesis that either mean(x1) = 0 or mean(x1) = mean(x2) with 80% power at alpha = 0.05
}
\examples{
minimumN(x1 = c(1, 2, 3, 4, 5), x2 = c(2, 3, 4, 5, 6))
}
