% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pap.R
\docType{data}
\name{pap}
\alias{pap}
\title{Pride and Prejudice
It is used to demo of this package.
The book Pride and Prejudice is a novel by Jane Austen published in 1813.}
\source{
\url{https://www.gutenberg.org/ebooks/1342}
}
\description{
Pride and Prejudice
It is used to demo of this package.
The book Pride and Prejudice is a novel by Jane Austen published in 1813.
}
