% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Takes data x and number of principle component npc and approximate x based on npc.}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A dataframe}

\item{npc}{number of principle component}
}
\value{
approximation
}
\description{
Takes data x and number of principle component npc and approximate x based on npc.
}
