% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plot.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{create plots for go and kegg}
\usage{
plot_enrichment(
  enrich_result,
  type = c("dotplot", "cnetplot"),
  showCategory = 10
)
}
\arguments{
\item{enrich_result}{An enrichment result object .}

\item{type}{Character. Plot type: `"dotplot"` or `"cnetplot"` (default: `"dotplot"`).}

\item{showCategory}{Integer. Number of categories to display (default: 10).}
}
\value{
A dotplot  or a plot produced by cnetplot()
}
\description{
Creates a visualisation of an enrichment result using either a dotplot or a cnetplot.
}
