% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{predict.logistic}
\alias{predict.logistic}
\title{FILL ME IN}
\usage{
\method{predict}{logistic}(fit, X, prob_cut = 0.5)
}
\arguments{
\item{fit}{FILL ME IN}
}
\value{
FILL ME IN
}
\description{
FILL ME IN
}
