% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichtment_kegg.R
\name{run_enrichment_go_kegg}
\alias{run_enrichment_go_kegg}
\title{Enrichemt script for kegg}
\usage{
run_enrichment_go_kegg(
  deg_table,
  use_only_pass = TRUE,
  kegg_organism = "hsa",
  show_top = 10
)
}
\arguments{
\item{deg_table}{A data frame containing at least `gene  and, if `use_only_pass = TRUE`, a logical `pass` column.}

\item{use_only_pass}{Logical. If TRUE, only genes with `deg_table$pass == TRUE` are used.}

\item{kegg_organism}{KEGG organism code.}

\item{show_top}{Number of top categories to show in downstream plotting}
}
\value{
A list with Go enrichment results, kegg enrichment results, id maps from SYMBOL to ENTREZID, Show_top for top 10 value from kegg
}
\description{
Takes a differential expression table  selects the  genes
runs GO  enrichment and KEGG pathway enrichment, and returns the results
}
