% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{scale2sd}
\alias{scale2sd}
\title{scale data vector}
\usage{
scale2sd(x)
}
\arguments{
\item{x}{a data vector}
}
\value{
scaled data vector
}
\description{
This function scale vector by meand and sd to normalized data with mean=0 and sd=1
}
