% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Converts the variables in a tibble data to small camel case}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with variables converted to small camel case
}
\description{
Converts the variables in a tibble data to small camel case
}
\examples{
data = data.frame(Item = c('Apple','oranges','pears'), price = c(1,2,3))
standardizeNames(data)
}
