% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Takes a numerical vector status and a numerical vector time, and calculate and plot the survival curve S(t)}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{numerical vector}

\item{time}{numerical vector}
}
\value{
None
}
\description{
Takes a numerical vector status and a numerical vector time, and calculate and plot the survival curve S(t)
}
\examples{
survCurv(status, time)
}
