% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Takes a vector that has been through the scale function in r and reverse the scaling, if any.}
\usage{
unscale(x)
}
\arguments{
\item{x}{A vector that was passed through scale}
}
\value{
unscaled vector
}
\description{
Takes a vector that has been through the scale function in r and reverse the scaling, if any.
}
\examples{
unscale(scale(c(1,2,3,4,5)))
}
