% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcloud.R
\name{wcloud}
\alias{wcloud}
\title{Making a wordcloud graph.}
\usage{
wcloud(x)
}
\arguments{
\item{x}{The book file data name.}
}
\value{
A wordcloud.
}
\description{
We want to make a wordcloud for the 200 most frequent words in the book.
}
\author{
Yi Xu
}
