% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_hist.R
\name{wf_hist}
\alias{wf_hist}
\title{Histogram for frequency}
\usage{
wf_hist(x)
}
\arguments{
\item{x}{book file name}
}
\value{
10 most frequent words
}
\description{
We want to make a histogram for the 10 most frequent words in the book.
}
\author{
Yi Xu
}
