% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{glm.logit.ridge}
\alias{glm.logit.ridge}
\title{ridge logistic regression using design matrix and Y}
\usage{
glm.logit.ridge(X, y, lambda = 1, tol = 1e-07, maxit = 50)
}
\arguments{
\item{X}{FILL ME IN}
}
\value{
FILL ME IN
}
\description{
FILL ME IN
}
