% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{logLik}
\alias{logLik}
\title{log likelihood function for logistic regression}
\usage{
logLik(eta, y)
}
\arguments{
\item{eta}{FILL ME IN}

\item{y}{FILL ME IN}
}
\value{
FILL ME IN
}
\description{
FILL ME IN
}
